# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reckora API",
    "description": "HTTP API for the Reckora OSINT investigation engine.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/investigations": {
      "post": {
        "tags": [
          "investigations"
        ],
        "summary": "Create Investigation",
        "description": "Run a Reckora investigation and persist the result.\n\nMirrors ``reckora investigate --save`` from the CLI: ``--archive``,\n``--screenshot`` and ``--ai`` map onto request fields, the saved dossier\nis returned in full so the client doesn't need a follow-up GET. The\ninvoking user is recorded as the dossier's owner (Phase 5 RBAC).",
        "operationId": "create_investigation_api_v1_investigations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedDossierPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/subjects": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "List Subjects",
        "description": "List saved dossiers visible to the current user, newest first.\n\n- **Admins** see every saved dossier (including legacy un-owned rows\n  created by the CLI before RBAC).\n- **Viewers** see only dossiers they own or that have been explicitly\n  shared with them via the sharing endpoints.",
        "operationId": "list_subjects_api_v1_subjects_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubjectSummary"
                  },
                  "title": "Response List Subjects Api V1 Subjects Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subjects/{subject_id}": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Get Subject",
        "description": "Return the rehydrated dossier for ``subject_id`` (auth-gated).",
        "operationId": "get_subject_api_v1_subjects__subject_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedDossierPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "investigations"
        ],
        "summary": "Delete Subject",
        "description": "Delete a saved dossier.\n\nAuthorisation:\n\n- Admins can delete any subject (including legacy un-owned rows).\n- Owners can delete their own subjects.\n- Shared viewers cannot delete; they get 403 (the subject *is*\n  visible to them, so 404 would be misleading).\n- Non-owner / non-shared viewers get 404 to avoid leaking existence.",
        "operationId": "delete_subject_api_v1_subjects__subject_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "description": "shared viewer cannot delete a dossier they don't own"
          },
          "404": {
            "description": "subject not found or not visible to actor"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subjects/{subject_id}/dossier": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Get Subject Dossier",
        "description": "Render a saved dossier in the requested format (auth-gated).",
        "operationId": "get_subject_dossier_api_v1_subjects__subject_id__dossier_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(md|json|html|pdf)$",
              "default": "html",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/html": {},
              "text/markdown": {},
              "application/pdf": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IdentifierIn": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "Identifier type, e.g. username, domain, url, email."
          },
          "value": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Value"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "value"
        ],
        "title": "IdentifierIn",
        "description": "An identifier the client wants the orchestrator to consider."
      },
...
      "InvestigationRequest": {
        "properties": {
          "seed": {
            "$ref": "#/components/schemas/IdentifierIn"
          },
          "extras": {
            "items": {
              "$ref": "#/components/schemas/IdentifierIn"
            },
            "type": "array",
            "title": "Extras"
          },
          "archive": {
            "type": "boolean",
            "title": "Archive",
            "description": "Mint a Wayback Machine snapshot per evidence URL (best-effort, slow).",
            "default": false
          },
          "screenshot": {
            "type": "boolean",
            "title": "Screenshot",
            "description": "Capture a forensic PNG of each evidence URL via headless Chromium (requires the 'screenshots' extra; off by default).",
            "default": false
          },
          "ai": {
            "type": "boolean",
            "title": "Ai",
            "description": "Run the LLM reasoning layer (summary + hypotheses).",
            "default": false
          },
          "ai_iterations": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Ai Iterations",
            "description": "Number of recursive AgentLoop rounds. 0 (default) = passive summary only. >=1 lets the LLM propose follow-up identifiers, verify them, and re-correlate, expanding the dossier graph.",
            "default": 0
          },
          "ai_tools": {
            "type": "boolean",
            "title": "Ai Tools",
            "description": "When ai_iterations >= 1, allow the AgentLoop's LLM to call web_search and fetch_url so it can gather evidence beyond what the rule-based collectors found. Works under both OPENAI_API_KEY (chat-completions function calling) and ChatGPT OAuth (Responses-API function calling).",
            "default": false
          },
          "ai_tool_calls": {
            "type": "integer",
            "maximum": 32.0,
            "minimum": 1.0,
            "title": "Ai Tool Calls",
            "description": "Per-iteration tool-call budget when ai_tools is true.",
            "default": 8
          },
          "breach": {
            "type": "boolean",
            "title": "Breach",
            "description": "Enable the data-leak surface: Have I Been Pwned breach lookup for email identifiers (requires HIBP_API_KEY) plus a public document-share / paste-site probe (Scribd, pdfcoffee, pdfslide, SlideShare, Issuu, 4shared, archive.org, Pastebin) for username + email seeds. Off by default.",
            "default": false
          },
          "anchor": {
            "type": "boolean",
            "title": "Anchor",
            "description": "Compute a cross-trace Merkle root and submit it to public OpenTimestamps calendars for tamper-evident timestamping (requires network access to the calendar fleet; off by default).",
            "default": false
          },
          "llm_provider": {
            "type": "string",
            "enum": [
              "auto",
              "openai",
              "chatgpt_oauth",
              "agentrouter"
            ],
            "title": "Llm Provider",
            "description": "LLM backend used when ai=true. 'auto' (default) tries OPENAI_API_KEY first, then ChatGPT OAuth. 'openai' / 'chatgpt_oauth' / 'agentrouter' pin the request to one path. The AgentRouter path uses the per-user BYOK key saved on the account, falling back to AGENTROUTER_API_KEY.",
            "default": "auto"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "seed"
        ],
        "title": "InvestigationRequest",
        "description": "Body for ``POST /api/v1/investigations``."
      },
...
      "SavedDossierPayload": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "subject": {
            "additionalProperties": true,
            "type": "object",
            "title": "Subject"
          },
          "traces": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Traces"
          },
          "timeline": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Timeline"
          },
          "anomalies": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Anomalies"
          },
          "edges": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Edges"
          },
          "ai": {
            "additionalProperties": true,
            "type": "object",
            "title": "Ai"
          },
          "anchor": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchor"
          },
          "owner_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Username"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "created_at",
          "subject",
          "traces",
          "timeline",
          "anomalies",
          "edges",
          "ai"
        ],
        "title": "SavedDossierPayload",
        "description": "Full dossier returned by ``GET /api/v1/subjects/{id}``.\n\nShape matches :func:`reckora.reports.json_export.to_dossier_dict` so the\nfrontend can reuse the same TypeScript type for both endpoints.\n``owner_username`` is omitted (``None``) for un-owned dossiers; see\n:class:`SubjectSummary` for the rationale."
      },
...
      "SubjectSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "seed": {
            "$ref": "#/components/schemas/IdentifierIn"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "identifier_count": {
            "type": "integer",
            "title": "Identifier Count"
          },
          "trace_count": {
            "type": "integer",
            "title": "Trace Count"
          },
          "edge_count": {
            "type": "integer",
            "title": "Edge Count"
          },
          "has_summary": {
            "type": "boolean",
            "title": "Has Summary"
          },
          "has_hypotheses": {
            "type": "boolean",
            "title": "Has Hypotheses"
          },
          "has_anchor": {
            "type": "boolean",
            "title": "Has Anchor",
            "default": false
          },
          "owner_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Username"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "seed",
          "created_at",
          "identifier_count",
          "trace_count",
          "edge_count",
          "has_summary",
          "has_hypotheses"
        ],
        "title": "SubjectSummary",
        "description": "List-row payload returned by ``GET /api/v1/subjects``.\n\n``owner_username`` is ``None`` for legacy un-owned dossiers (created\nvia the CLI before RBAC landed, or before being claimed by an\nadmin). The frontend renders these as \"system\" rows."
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        }
      }
    }
  }
}